+++ src/main/kotlin/com/github/sleepypanda/feesh/settings/categories/Overlays.kt
@@ -104,8 +104,8 @@ object Overlays : CategoryKt("Overlays") {
 
     var trackersAutoPauseSeconds by int(180) {
         this.name = Translated("Idle seconds to auto-pause all trackers")
-        this.description = Translated("Pauses elapsed timers on various widgets when you stop fishing for this long (in seconds).\n${YELLOW}Make sure to choose enough time to kill any fishing mob and get loot from it, so it gets counted by the trackers!")
-        this.range = 30..300
+        this.description = Translated("Pauses elapsed timers on various widgets when you stop fishing for this long (in seconds).\n${YELLOW}Make sure to choose enough time to kill any fishing mob and get loot from it, so it gets counted by the trackers before pause!")
+        this.range = 10..300
         this.slider = true
     }