+++ docs/dev/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 -
 
+## Bugfixes
+
+- Decreased minimum value to 10 seconds for "Idle seconds to auto-pause all trackers" setting.
+
 # 1.7.0
 
 Released on: 2026-05-26