+++ docs/dev/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added Flipflopper and Seashine creatures to HP tracker and mob highlighting.
 - Added Flipflopper Shard and Seashine Shard to Fishing profit tracker.
 - Decreased minimum value to 10 seconds for "Idle seconds to auto-pause all trackers" setting.
+- Added on-screen alert when reaching PB.
 
 # 1.7.0